fix(api/swagger): add the references as array to the detail endpoints
The swagger generator uses the serializer to generate the output, but it
only looks at the `list` serialization by default. The `references` in
the SicprodSerializer where only added for the `retrieve` action,
though.
We can refactor that and use a separate ListSerializer for all the
classes, this differentiation is then also picked up the the swagger
generator. To make things a bit easier to read, we move the `references`
field setup to a `ReferencesMixin` which then gets reused by the
individual serializers.